done.
This project is created by Aysu Şen as part of learning Android development with Kotlin.
It is a fun mini game to practice:

Using View Binding

Implementing CountDownTimer and Handler

Working with ImageViews and GridLayouts

Writing clean and understandable Kotlin code

I developed this project to improve my Android skills and understand how to build interactive games with simple logic.